refactor: insert Name to Car
- Car객체에 Name객체가 포함되도록 수정
- Name 클래스 삭제
- 관련 테스트 재진행
